demo.py
from langchain import OpenAI, SQLDatabase, SQLDatabaseChain
from langchain.embeddings import HuggingFaceEmbeddings
from langchain.vectorstores import FAISS
from langchain.document_loaders import TextLoader
from langchain.text_splitter import CharacterTextSplitter
from langchain.chains import RetrievalQA
import mysql.connector
import pandas as pd

# MySQL数据库连接
db = SQLDatabase.from_uri("mysql+mysqlconnector://username:password@host/database")

# RAG知识库
embeddings = HuggingFaceEmbeddings()
docs = TextLoader('path/to/your/document.txt').load()
text_splitter = CharacterTextSplitter(chunk_size=1000, chunk_overlap=0)
documents = text_splitter.split_documents(docs)
vectorstore = FAISS.from_documents(documents, embeddings)

# 对话模型
llm = OpenAI(temperature=0)

# 基于MySQL的对话链
db_chain = SQLDatabaseChain(llm=llm, database=db, verbose=True)

# 基于RAG的对话链
qa_chain = RetrievalQA.from_chain_type(llm=llm, chain_type="stuff", retriever=vectorstore.as_retriever())

# 用户对话数据更新函数
def update_user_data(user_id, query, response):
    # 实现将用户对话数据更新到MySQL数据库的逻辑

# 数据筛选和清洗函数
def process_user_data(filter_criteria):
    # 实现从MySQL数据库筛选数据的逻辑
    # 实现数据清洗的逻辑
    # 生成用于微调的数据集格式

# 主对话循环
def chat():
    while True:
        user_input = input("用户: ")
        if user_input.lower() == 'quit':
            break
        
        # 判断是否为数据库相关查询
        if "数据库" in user_input:
            response = db_chain.run(user_input)
        else:
            response = qa_chain.run(user_input)
        
        print("AI: ", response)
        
        # 更新用户对话数据
        update_user_data(user_id, user_input, response)

if __name__ == "__main__":
    chat()