requirements.txt
langchain
openai
mysql-connector-python
faiss-cpu
pandas
numpy
sentence-transformers
flask
sqlalchemy
pymysql
python-dotenv